requirements.txt
beautifulsoup4
pydantic-settings
python-dateutil
python-telegram-bot[job-queue]
requests
types-beautifulsoup4
